+++ src/components/common/HeadingText.tsx
